rex ui test deduplication
